fix(T11983): relocate whoami summary helper to core (CLI boundary lint)
Move `_printWhoamiSummaryAndOfferTui` (38 LOC) from
`packages/cleo/src/cli/commands/setup.ts` into a new
`packages/core/src/setup/whoami-summary.ts` module and export it
through `packages/core/src/setup/index.ts` as `printWhoamiSummaryAndOfferTui`.

Import it back in the CLI command and update the e2e test to import
from `@cleocode/core/setup` instead of `../setup.js`.

Lint: `node scripts/lint-cli-package-boundary.mjs --check` → PASS 33/33 (baseline, no net-new).
Tests: setup-wizard-e2e.test.ts + setup-command.test.ts → 61 passed.
